balls now work on floats when it comes to speed and position so integer div does not ruin collisons

+++ CrazyBallz/Data/IBall.cs
@@ -5,17 +5,17 @@ namespace Data
 {
     public abstract class IBall
     {
-        public abstract int Position_X { get; set; }
-        public abstract int Position_Y { get; set; }
+        public abstract float Position_X { get; set; }
+        public abstract float Position_Y { get; set; }
         public abstract int Radius { get; }
-        public abstract int Speed_X { get; set; }
-        public abstract int Speed_Y { get; set; }
+        public abstract float Speed_X { get; set; }
+        public abstract float Speed_Y { get; set; }
         public abstract int Mass { get; set; }
         public abstract Vector2 BoardSize { get; set; }
 
         public abstract event PropertyChangedEventHandler? PropertyChanged;
 
-        public static IBall CreateBall(int postion_X, int postion_Y, int radius, int speed_X, int speed_Y, int mass)
+        public static IBall CreateBall(float postion_X, float postion_Y, int radius, float speed_X, float speed_Y, int mass)
         {
             return new Ball(postion_X, postion_Y, radius, speed_X, speed_Y, mass);
         }
