+++ CrazyBallz/Logic/BallManager.cs
@@ -123,7 +123,7 @@ namespace Logic
 
         public void HandleCollision(IBall a, IBall b)
         {
-            int Vx1, Vy1, Vx2, Vy2;
+            float Vx1, Vy1, Vx2, Vy2;
 
             Vx1 = (a.Mass * a.Speed_X + b.Mass * b.Speed_X - b.Mass * (a.Speed_X - b.Speed_X)) / (a.Mass + b.Mass);
             Vy1 = (a.Mass * a.Speed_Y + b.Mass * b.Speed_Y - b.Mass * (a.Speed_Y - b.Speed_Y)) / (a.Mass + b.Mass);
